Move state and scene files.
